requirements.txt
Django>=3.0.8,<3.1.0
Djangorestframework>=3.11.0,<3.12.0
djangorestframework-jwt>= 1.11.0,<1.12.0
djangorestframework-simplejwt==4.2.0
django-environ==0.4.5

